requirements.txt
requests
cryptography
Flask
Flask-Cors
SQLAlchemy
psycopg2-binary
gunicorn
python-dotenv